todo-backend/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.7
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Mako==1.3.8
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.34.0
